MAT-1714 Moved vsac model into a non nested package for GWT.
