+++ src/main/java/mat/Bonnie.gwt.xml
@@ -44,6 +44,6 @@
     <source path='mat.shared.diagramObject'/>
     <source path='model'/>
     <source path='dto'/>
-    <source path='vsac'/>
+    <source path='vsacmodel'/>
     <!-- <add-linker name="xsiframe" /> -->
 </module>
