+++ src/main/java/mat/Mat.gwt.xml
@@ -47,5 +47,5 @@
     <source path='mat.shared.diagramObject'/>
     <source path='model'/>
     <source path='dto'/>
-    <source path='vsac'/>
+    <source path='vsacmodel'/>
 </module>
