+++ src/main/java/mat/client/clause/QDMAvailableValueSetPresenter.java
@@ -30,7 +30,7 @@ import mat.model.QualityDataSetDTO;
 import mat.model.MatValueSetTransferObject;
 import mat.shared.ConstantMessages;
 import org.gwtbootstrap3.client.ui.Button;
-import mat.vsac.model.ValueSet;
+import mat.vsacmodel.ValueSet;
 
 import java.util.ArrayList;
 import java.util.Collections;
