+++ src/main/java/mat/client/cqlworkspace/AbstractCQLWorkspacePresenter.java
@@ -34,7 +34,7 @@ import mat.dto.VSACCodeSystemDTO;
 import mat.model.CQLValueSetTransferObject;
 import mat.model.CodeListSearchDTO;
 import mat.model.GlobalCopyPasteObject;
-import mat.vsac.model.ValueSet;
+import mat.vsacmodel.ValueSet;
 import mat.model.cql.CQLCode;
 import mat.model.cql.CQLDefinition;
 import mat.model.cql.CQLFunctionArgument;
