+++ src/main/java/mat/server/service/impl/CodeListXLSGenerator.java
@@ -7,7 +7,7 @@ import mat.model.clause.Measure;
 import org.apache.poi.hssf.usermodel.HSSFCellStyle;
 import org.apache.poi.hssf.usermodel.HSSFSheet;
 import org.apache.poi.hssf.usermodel.HSSFWorkbook;
-import mat.vsac.model.ValueSet;
+import mat.vsacmodel.ValueSet;
 
 import java.sql.Timestamp;
 import java.util.ArrayList;
