+++ src/main/java/mat/server/service/impl/MeasurePackageServiceImpl.java
@@ -60,7 +60,7 @@ import org.apache.commons.logging.Log;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
-import mat.vsac.model.ValueSet;
+import mat.vsacmodel.ValueSet;
 
 import java.io.IOException;
 import java.util.ArrayList;
