+++ src/main/java/mat/server/service/impl/SimpleEMeasureServiceImpl.java
@@ -57,7 +57,7 @@ import org.w3c.dom.Document;
 import org.w3c.dom.Node;
 import org.w3c.dom.NodeList;
 import org.xml.sax.InputSource;
-import mat.vsac.model.ValueSet;
+import mat.vsacmodel.ValueSet;
 
 import javax.xml.parsers.DocumentBuilder;
 import javax.xml.parsers.DocumentBuilderFactory;
