+++ src/main/java/mat/server/service/impl/ValueSetXLSGenerator.java
@@ -5,7 +5,7 @@ import mat.model.ListObject;
 import org.apache.poi.hssf.usermodel.HSSFCellStyle;
 import org.apache.poi.hssf.usermodel.HSSFSheet;
 import org.apache.poi.hssf.usermodel.HSSFWorkbook;
-import mat.vsac.model.ValueSet;
+import mat.vsacmodel.ValueSet;
 
 import java.sql.Timestamp;
