+++ src/main/java/mat/vsac/VsacService.java
@@ -4,11 +4,11 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.json.JsonMapper;
-import mat.vsac.model.BasicResponse;
-import mat.vsac.model.CodeSystemVersionResponse;
-import mat.vsac.model.ValueSetResult;
-import mat.vsac.model.ValueSetWrapper;
-import mat.vsac.model.VsacCode;
+import mat.vsacmodel.BasicResponse;
+import mat.vsacmodel.CodeSystemVersionResponse;
+import mat.vsacmodel.ValueSetResult;
+import mat.vsacmodel.ValueSetWrapper;
+import mat.vsacmodel.VsacCode;
 import mat.vsac.util.VsacConverter;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.collections4.CollectionUtils;
