+++ src/main/java/mat/vsacmodel/CodeSystemVersionResponse.java
@@ -1,4 +1,4 @@
-package mat.vsac.model;
+package mat.vsacmodel;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
