+++ src/main/resources/MAT.properties
@@ -32,7 +32,7 @@ mat.qdm.default.expansion.id=Most Recent Code System Versions in VSAC
 
 mat.codesystem.valueset.simultaneous.validations=5
 
-#mat.mat.vsac.server.url = https://vsac.nlm.nih.gov/vsac
+#mat.mat.vsacmodel.server.url = https://vsac.nlm.nih.gov/vsac
 
 #MAT 2 factor authentication properties.
 2FA_AUTH_CLASS=mat.server.twofactorauth.DefaultOTPValidatorForUser
